sql_agent_project/requirements.txt
fastapi
uvicorn
psycopg2
langchain
langchain-openai
python-dotenv
